receiving request from SparqlBlocks.
i.e when the request comes from sparqlblocks about generating HTML+SVG page, this app.js file receives a request and send those request to Routes in order to generate HTML+SVG page with export of query and results. and from the same page we have lot request sent to app.js file. i.e in order to download a  query (textual and visual) and results in JSON file and one more request for History of same query with differet results.